rag_pipeline/requirements.txt
flask
requests
beautifulsoup4
sentence-transformers
faiss-cpu
openai
